services/fraud-detection-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
joblib==1.4.2
imbalanced-learn==0.12.4
confluent-kafka==2.5.3
pytest==8.3.3
httpx==0.27.2
pydantic-settings==2.5.2
prometheus-fastapi-instrumentator==7.0.0
